requirements.txt
crewai
langchain
langchain_openai
python-dotenv
requests
